+++ src/block.cpp
@@ -3,8 +3,8 @@
 #include <raylib.h>
 #include <random>
 
-int Block::width;
-int Block::height;
+float Block::width;
+float Block::height;
 
 Block::Block(Point upLeftPos, bool spawnsBall)
     : upLeftPos(upLeftPos), spawnsBall(spawnsBall)
